fer_service/requirements.txt
opencv-python
numpy
tensorflow
deepface
flask
flask-cors
requests
twilio
python-dotenv
